Exercicio Pronto
Exercício feito.

+++ Exercícios Java/Exercícios CAP 10/Exercicio de Fixacao - Lista de Funcionarios/src/entities/Employee.java
@@ -31,7 +31,7 @@ public class Employee {
         return salary;
     }
 
-    public void salaryIncrease (int x){
+    public void salaryIncrease (float x){
         salary *= (1+x/100);
     }
     public String toString(){
